Support passing in a credential-helper-cmd
We have a situation where we assume a role in another account from where
the command is run, and want to allow using credentials for said account.

We can't use assumption directly in C7N - There are situations where we want
our additional capabilitys - and passing in credentials as environment
variables doesn't include refreshability.

This adds a new binary argument, which specifies a commnand to invoke
- allowing us to implement custom, refreshable auth using a command line
program.