Fix an import conflict when using skip-validation as an option
Refactoring in Jan 2020 moved this around, and lead to a situation
where importing schema validation would overwrite the variable
used to control validation.

Generally this isn't an issue, but in our run mode we have policies files
with 6000+ items, and per-policy validation can take over an hour in said
situation.

This restores the flags value by using an aliased import instead, and
makes it all hunky dory.